добавил webpack и webpac-dev-server, разбил код на модули
